+++ CHANGELOG.md
@@ -18,6 +18,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - Added `ToCigletIterator` trait, implemented on structs which can be converted to an iterator of ciglets
 - Added `LenInAlignment` trait, replacing the previous `match_len` and offering a more flexible API
 - Added `VARIANTS` constants to pHMM state enums (behind `dev-phmm` feature gate)
+- Added index comparison methods to `PhmmIndex` (behind `dev-phmm` feature gate)
 
 ### Changed